+++ PyFiSync/__init__.py
@@ -2,5 +2,6 @@
 # -*- coding: utf-8 -*-
 import sys
 sys.dont_write_bytecode = True
+
 from .main import __version__,__author__ 
 from .main import cli